requirements.txt
about-time==4.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alive-progress==3.3.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
autograd==1.8.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
clarabel==0.11.1
click==8.3.1
cma==4.4.1
comocma>=0.6.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cvxpy==1.7.5
cycler==0.12.1
datasets==4.4.2
debugpy==1.8.19
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.0
ecos==2.0.14
executing==2.2.1
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
graphemeu==0.7.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline==0.2.1
moocore==0.1.10
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
osqp==1.0.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymoo==0.6.1.6
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qpsolvers==4.8.2
quadprog==0.1.13
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
scs==3.2.9
sentry-sdk==2.48.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
torchjd==0.10.0
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wandb==0.23.1
wcwidth==0.2.14
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
